Add another (more naïve) solution to 2019/15.
Based on an earlier attempt. Turned out that following the left wall
from start to goal revealed enough of the map to give the correct
answer. However, following the right wall instead did not.
This extra solution first follows the left wall and then restarts and
follows the right wall. This reveals the entire map for my input.

I mainly saved this because I liked the follow left/right wall
implementation that I wrote, before scrapping it and doing a DFS instead.