2020/20 completed.
Part 2 took me a while. I confused myself a lot with rotating and
flipping the tiles, mostly because of silly initial implementation of
the border to tile lookup. In the end, keeping rotation separate from
flipping solved my confusion.

Then, my monster finding didn't initially work for multiple monsters on
the same line...

All in all: Urgh. (But still fun, in the end.)