2020/22 completed.
My initial solution for part 2 took about a minute to run. Partly
because I didn't check for seen states in each players' hands
separately, only the whole game state, but even after changing that it
was slow. I ripped it all out and restarted from scratch, and now it
runs in a few seconds. ¯\_(ツ)_/¯