2018/12 improved.
My initial idea was to use (and practice) bit masking for calculating
the states, but after realizing that the used pots could go below 0, I
dismissed that idea after implementing it halfway. Later, I realized I
could keep track of the minimum ID separately and still use bit masks.
This is the result, a bit faster than the previous version.