2018/24 completed.
The infinite battle when boost is one lower than the minimum needed for
victory was evil. Had I not been using binary search to find the
minimum, I would probably have gotten a bit stumped there, but now I saw
that both the boost values below and above that worked fine before it
happened the first time. I (correctly) guessed that the value above the
infinitely running one was the answer, before debugging the issue. :)