Faster solution to 2021/18.
My initial idea was to store the inputs as flat lists of number, depth
pairs. I spent some time on that implementation before scrapping it and
going for a recursive solution instead (all the simple examples worked,
but something broke in the first reduction of the "slightly larger"
example and I couldn't figure out what).

After seeing a successful flat implementation in the subreddit, I
decided to give it another try (from scratch, as my original attempt is
gone). It ended up being slightly faster than the recursive solution.