Improve 2022/19.
Cut the runtime in ~half by not allowing building a bot that you could
have built last minute but didn't. It never makes sense to skip building
something you can afford and build it the next minute instead.
This even makes the dubious "if you can build a geode robot, always do
so" optimization obsolete, which is good, as that is not guaranteed to
be correct.