Fix 2022/19.
Found an input where the optimization of never skipping a robot build if
we had the max number of ores needed for any recipe in stock, so had to
remove that to be correct in the general case. This increased the
runtime by about 4x, so brought back Multicore to compensate.

This _should_ now be correct for any input, as the rest of the
optimizations can be proven to be correct, if I'm not forgetting
something...