2022/22 done.
I was slow today. In part 1, it took a while to get the wrapping right,
and in part 2, I debugged my wrapping step by step for a long time
before realizing I changed direction due to wrapping even if I hit a
wall on the other side, which obviously wasn't correct.

Don't like that part 2 has a hardcoded cube layout. We'll see if I have
enough willpower to go back and make a generic solution later...