Speed up 2022/24.
The normal trick of using bitmasked integers instead of Arrays for
storing position helped a lot. Map storage was changed to a sparse list
of walls, and by adding an extra wall behind the start and beyond the
goal, a lot of special checks could be removed. Lastly, using Hash
instead of Set saved a few ms.