Add alternate solution to 2022/24.
I was initially going for a more "classic" BFS with a normal queue etc,
but scrapped that when it took forever to run. First after solving the
puzzle with the new solution, I realized I had "forgotten" the set of
visited positions (I somehow thought including the time in the queued
state would make that obsolete, I blame sleep-deprivation).
At that point, curiosity took hold of me, and I wanted to see how much
worse it would perform, so here it is. The answer is: slightly slower,
but not by much.