Minor optimization of 2022/24 (classicbfs solution).
It's obvious the blizzard pattern repeats with a cycle of
lcm(width, height) (700 for my input), so we can simply reuse the
already calculated states. This only affects the "classic BFS" solution,
since the main (faster one) doesn't keep track of the blizzard positions
over time (and adding that makes it slower, I tried).