Further speed up of 2025/09.
Cheat a bit in the rectangle checking logic. By only checking for
intersections with the polygon, the check is much faster. It's not
correct in the general sense, but correct enough for this problem.

Include a test input which showcases the issue. Keep the old version as
movie_theater_complete.rb.